updated folder structure and initialised pages
