+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/break-down-interest/break-down-interest-page.ts
