+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/claim-interest-options/claim-interest-options-page.ts
@@ -6,9 +6,7 @@ import { radioButtons, subheadings } from './claim-interest-options-content';
 @AllMethodsStep()
 export default class ClaimInterestOptionsPage extends ExuiPage(BasePage) {
   async verifyContent() {
-    await super.runVerifications([
-      super.expectText(subheadings.claimInterestOptions),
-    ]);
+    await super.runVerifications([super.expectText(subheadings.claimInterestOptions)]);
   }
 
   async selectSameRateInterest() {
