+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/interest-claim-from/interest-claim-from-content.ts
