+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/interest-claim-until/interest-claim-until-content.ts
