+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/interest-from-specific-date/interest-from-specific-date-content.ts
