+++ playwright-e2e/pages/exui/solicitor-events/create-claim/lr-spec/same-rate-interest-selection/same-rate-interest-selection-content.ts
